requirements.txt
torch>=1.8
torchvision>=0.9
numpy>=1.21
pillow>=9.0.0
opencv-python>=4.7.0
matplotlib>=3.5.0
scipy>=1.8.0
scikit-image>=0.19.0
flask>=2.2.0

